加入place_object2,3_tag和remove_object1,2_tag以及filter读取框架，并整理代码

+++ swf_reader/Shapes/Records/style_change_shape_record.h
@@ -20,8 +20,8 @@ namespace swf_reader::shapes::records
         u32 fillstyle0;
         u32 fillstyle1;
         u32 linestyle;
-        i16 move_delta_x;
-        i16 move_delta_y;
+        i32 move_delta_x;
+        i32 move_delta_y;
 
         inline ShapeRecordType get_type() const override { return ShapeRecordType::StyleChangeRecord; }
